feat: AppBar 추가 및 타이틀, 액션 버튼 기능 구현
