+++ lib/main.dart
@@ -2,7 +2,7 @@ import 'package:damdiet/routes/app_routes.dart';
 import 'package:damdiet/screen/community/community_detail_screen.dart';
 import 'package:damdiet/screen/community/community_write_screen.dart';
 import 'package:damdiet/screen/community/kcal_calculator_screen.dart';
-import 'package:damdiet/screen/home_screen.dart';
+import 'package:damdiet/screen/home/home_screen.dart';
 import 'package:damdiet/screen/mypage/mypage_favorite_products_screen.dart';
 import 'package:damdiet/screen/mypage/mypage_my_community_screen.dart';
 import 'package:damdiet/screen/mypage/mypage_my_orders_screen.dart';
@@ -12,7 +12,6 @@ import 'package:damdiet/screen/payment_screen.dart';
 import 'package:damdiet/screen/search/product_detail_screen.dart';
 import 'package:damdiet/screen/search/products_screen.dart';
 import 'package:flutter/material.dart';
-import 'package:provider/provider.dart';
 
 void main() {
   runApp(const DamDietApp());
